requirements.txt
python-dotenv>=1.0.1
pymongo>=4.7.0
pydantic>=2.7.0
requests>=2.32.0
youtube-transcript-api>=0.6.2
google-api-python-client>=2.142.0
openai>=1.42.0
streamlit>=1.37.0
numpy>=1.26.0
pandas>=2.0.0
plotly>=5.22.0
voyageai>=0.3.5


langchain-community>=0.3.0
langchain_openai
langchain>=0.2.14
tiktoken>=0.7.0
langchain-experimental>=0.0.59
llmlingua>=0.1.14
pytube>=15.0.0
scikit-learn>=1.4.0
networkx>=3.0
graspologic>=3.0
pyyaml>=6.0
unidecode>=1.3.0
jellyfish>=0.12.0
